+++ CHANGELOG.md
@@ -32,6 +32,7 @@ NOTE: please use them in this order.
 - Dropped support for sphinx 5 ([#110](https://github.com/rstcheck/rstcheck-core/pull/110))
 - Added python 3.13 to CI test pool ([#110](https://github.com/rstcheck/rstcheck-core/pull/110))
 - Add sphinx 8 to test pool for python version > 3.9 ([#111](https://github.com/rstcheck/rstcheck-core/pull/111))
+- Added python 3.14 to tox config and CI test pool ([#113](https://github.com/rstcheck/rstcheck-core/pull/113))
 
 ## [v1.2.1 (2024-03-23)](https://github.com/rstcheck/rstcheck-core/releases/v1.2.1)