+++ CHANGELOG.md
@@ -35,6 +35,7 @@ NOTE: please use them in this order.
 - Added python 3.14 to tox config and CI test pool ([#113](https://github.com/rstcheck/rstcheck-core/pull/113))
 - Dropped python 3.9 support ([#115](https://github.com/rstcheck/rstcheck-core/pull/115))
 - Add sphinx v9 to test runner ([#116](https://github.com/rstcheck/rstcheck-core/pull/116))
+- Upgrade base python version to v3.14 for automations ([#117](https://github.com/rstcheck/rstcheck-core/pull/117))
 
 ## [v1.2.1 (2024-03-23)](https://github.com/rstcheck/rstcheck-core/releases/v1.2.1)